Added ft_process_token_list_utils2.c
